Support step-based function counters and timers for step registries

+++ micrometer-core/src/main/java/io/micrometer/core/instrument/composite/CompositeFunctionTimer.java
@@ -41,7 +41,7 @@ class CompositeFunctionTimer<T> extends AbstractCompositeMeter<FunctionTimer> im
     }
 
     @Override
-    public long count() {
+    public double count() {
         return firstChild().count();
     }
